requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
openai
python-dotenv
pandas
ollama
pytesseract
pillow
fake_useragent
bs4
selenium 
webdriver-manager